refactor: eventlog is tunnel's internal
