+++ internal/eventlog/event.go
