+++ internal/eventlog/interfaces.go
