+++ internal/eventlog/manager.go
