+++ internal/manager/manager.go
@@ -3,7 +3,7 @@ package manager
 import (
 	"sync"
 
-	"github.com/Codename-Uranium/tunnel/eventlog"
+	"github.com/Codename-Uranium/tunnel/internal/eventlog"
 	"github.com/Codename-Uranium/tunnel/internal/runtime"
 	"github.com/Codename-Uranium/tunnel/internal/storage"
 	"github.com/Codename-Uranium/tunnel/internal/wireguard"
